ミスを修正

+++ kernel/pci.cpp
@@ -12,7 +12,7 @@ namespace {
     using namespace pci;
 
     /** @brief CONFIG_ADDRESS用の32ビット整数を生成する */
-    uint8_t MakeAddress(uint8_t bus, uint8_t device, uint8_t function, uint8_t reg_addr) {
+    uint32_t MakeAddress(uint8_t bus, uint8_t device, uint8_t function, uint8_t reg_addr) {
         auto shl = [](uint32_t x, unsigned int bits) {
             return x << bits;
         };
